Deduplicate writing null case of struct wrapped nullable pointers
